modified:   opencl/1_vecAdd/main.cc 	modified:   opencl/1_vecAdd/simple_add.cl.c

+++ opencl/1_vecAdd/simple_add.cl.c
@@ -1,6 +1,6 @@
-void kernel simple_add(global const int* A, 
-		       global const int* B, 
-		       global int* C)
+void kernel simple_add(global const float* A, 
+		       global const float* B, 
+		       global float* C)
 {
 	int i = get_global_id(0);
 	C[i] = A[i] + B[i];
